remove useless as u8

+++ rust-gameboy/src/joypad.rs
@@ -48,7 +48,7 @@ impl Joypad {
     }
 
     pub fn extract_interupt(&mut self) -> u8 {
-        let r = self.interupt as u8;
+        let r = self.interupt;
         self.interupt = 0;
         return r;
     }
